tests/functional/requirements.txt
pydantic==1.9.0
pytest==7.1.1
pytest-asyncio==0.18.3
aioredis==2.0.1
aiohttp==3.8.1
aiopg==1.3.3
psycopg2-binary==2.9.3
requests==2.27.1